Extract out container unittest
